Create real f(x)s and allow calls
This commit makes usage of the lambda-engine service instead the example nginx
image we were using before. For using the engine we need to set few environment
variables when we are creating the environment.

We are also introducing the endpoint for calling the functions without knowing
where the service is.